data-viewer/api/requirements.txt
PyYAML==6.0.1
requests==2.32.3
fastapi==0.111.1
uvicorn[standard]==0.30.1
pydantic-settings==2.3.4
google-cloud-logging==3.10.0
pyproj==3.6.1
wget==3.2
pandas==2.2.2
pyarrow==23.0.0
s3fs==2026.2.0
